create final edit (#15)

+++ src/modules/roomNonAvailability/roomNonAvailability.controller.ts
@@ -6,6 +6,7 @@ import {
   GetRoomNonAvailabilitiesByTenant,
   UpdateRoomNonAvailabilityDTO,
 } from "./dto/roomNonAvailability";
+import { ApiError } from "../../utils/api-error";
 
 export class RoomNonAvailabilityController {
   roomNonAvailabilityService: RoomNonAvailabilityService;
@@ -15,12 +16,15 @@ export class RoomNonAvailabilityController {
   }
 
   createRoomNonAvailability = async (req: Request, res: Response) => {
-    const tenantId = Number(res.locals.user.tenant.id);
+    const authUserId = Number(res.locals.user.id);
+    if (!authUserId) {
+      throw new ApiError("Unauthorized", 403);
+    };
     const roomId = Number(req.params.roomId);
     const data = plainToInstance(CreateRoomNonAvailabilityDTO, req.body);
     const result =
       await this.roomNonAvailabilityService.createRoomNonAvailability(
-        tenantId,
+        authUserId,
         roomId,
         data
       );
@@ -28,46 +32,58 @@ export class RoomNonAvailabilityController {
   };
 
   updateRoomNonAvailability = async (req: Request, res: Response) => {
-    const tenantId = Number(res.locals.user.tenant.id);
+    const authUserId = Number(res.locals.user.id);
+    if (!authUserId) {
+      throw new ApiError("Unauthorized", 403);
+    }
     const id = Number(req.params.id);
     const data = plainToInstance(UpdateRoomNonAvailabilityDTO, req.body);
     const result =
       await this.roomNonAvailabilityService.updateRoomNonAvailability(
         id,
-        tenantId,
+        authUserId,
         data
       );
     return res.status(200).send(result);
   };
 
   deleteroomNonAvailability = async (req: Request, res: Response) => {
-    const tenantId = Number(res.locals.user.tenant.id);
+    const authUserId = Number(res.locals.user.id);
+    if (!authUserId) {
+      throw new ApiError("Unauthorized", 403);
+    }
     const id = Number(req.params.id);
     const result =
       await this.roomNonAvailabilityService.deleteroomNonAvailability(
         id,
-        tenantId
+        authUserId
       );
     return res.status(200).send(result);
   };
 
   getAllRoomNonAvailabilitiesByTenant = async (req: Request, res: Response) => {
-    const tenantId = Number(res.locals.user.tenant.id);
+    const authUserId = Number(res.locals.user.id);
+    if (!authUserId) {
+      throw new ApiError("Unauthorized", 403);
+    }
     const query = plainToInstance(GetRoomNonAvailabilitiesByTenant, req.query);
     const result =
       await this.roomNonAvailabilityService.getAllRoomNonAvailabilitiesByTenant(
-        tenantId,
+        authUserId,
         query,
       );
     return res.status(200).send(result);
   };
 
   getAllRoomNonAvailabilitiesByRoom = async (req: Request, res: Response) => {
-    const tenantId = Number(res.locals.user.tenant.id);
+    const authUserId = Number(res.locals.user.id);
+    if (!authUserId) {
+      throw new ApiError("Unauthorized", 403);
+    }
     const roomId = Number(req.params.roomId);
     const result =
       await this.roomNonAvailabilityService.getAllRoomNonAvailabilitiesByRoom(
-        tenantId,
+        authUserId,
         roomId
       );
     return res.status(200).send(result);
